+++ R/tab_create_modify.R
@@ -1711,9 +1711,6 @@ tab_row_group <- function(
   # Capture the `rows` expression
   row_expr <- rlang::enquo(rows)
 
-  # Get the `stub_df` data frame from `data`
-  stub_df <- dt_stub_df_get(data = data)
-
   # Resolve the row numbers using `resolve_vars()`
   resolved_rows_idx <-
     resolve_rows_i(
@@ -1721,6 +1718,7 @@ tab_row_group <- function(
       data = data
     )
 
+  # Get the `stub_df` data frame from `data`
   stub_df <- dt_stub_df_get(data = data)
 
   # If the label is marked as HTML or Markdown and there's no `id` set