+++ NEWS.md
@@ -49,10 +49,15 @@
 
 * Fixed an issue where `tab_spanner_delim()` would fail to resolve a duplicate id (@olivroy, #1821).
 
-
 * Fixed an issue with multiple `text_replace()` calls would produce bad result with `cells_column_labels()` (@olivroy, #1824).
 
 * `tidyselect::where()`, `tidyselect::all_of()`, `tidyselect::any_of()` are now re-exported by gt.
+
+* `vec_fmt_markdown()` works correctly inside Quarto again (@olivroy, #1840).
+
+* Fixed a bug where spanners didn't render correctly in Quarto (#1839).
+
+* Improved footnote rendering in Quarto with `fmt_markdown()` (@olivroy, #1773)
 
 # gt 0.11.0